Fix underflow in calculating the modlist merge budget.
--HG--
branch : v3
extra : transplant_source : %84%5B%5Du%AFO%E2S%0C%97%17f%F3%29%88%F8%FBF%BB%C2

+++ kernel/castle_da.c
@@ -12165,7 +12165,7 @@ void castle_double_array_queue(c_bvec_t *c_bvec)
 int castle_double_array_init(void)
 {
     int ret, cpu, i, j;
-    int min_budget, budget;
+    unsigned long min_budget, budget;
 
     ret = -ENOMEM;
